Moved the 'make_gif' function over to the backend server's FIB workflow module and optimised input parameters required
