+++ src/murfey/server/api/workflow_fib.py
@@ -1,14 +1,19 @@
 import json
 import logging
+import os
 from importlib.metadata import entry_points
 from pathlib import Path
 
+import numpy as np
+import PIL.Image
 from fastapi import APIRouter, Depends
 from pydantic import BaseModel
-from sqlmodel import Session
+from sqlmodel import Session, select
 
+import murfey.util.db as MurfeyDB
 from murfey.server.api.auth import validate_instrument_token
 from murfey.server.murfey_db import murfey_db
+from murfey.util.config import get_machine_config
 from murfey.util.models import LamellaSiteInfo
 
 logger = logging.getLogger("murfey.server.api.workflow_fib")
@@ -57,3 +62,67 @@ def register_fib_milling_progress(
         "Received the following FIB metadata for registration:\n"
         f"{json.dumps(site_info.model_dump(exclude_none=True), indent=2, default=str)}"
     )
+
+
+class FIBGIFParameters(BaseModel):
+    lamella_number: int
+    images: list[Path]
+    output_file: Path
+
+
+@router.post("/sessions/{session_id}/make_gif")
+async def make_gif(
+    session_id: int,
+    gif_params: FIBGIFParameters,
+    db=murfey_db,
+):
+    # Load machine config and session info
+    session_entry = db.exec(
+        select(MurfeyDB.Session).where(MurfeyDB.Session.id == session_id)
+    ).one()
+    instrument_name = session_entry.instrument_name
+    visit_name = session_entry.visit
+    machine_config = get_machine_config(instrument_name=instrument_name)[
+        instrument_name
+    ]
+
+    # Create the directory structure
+    if not (output_dir := gif_params.output_file.parent).exists():
+        output_dir.mkdir(parents=True)
+        logger.debug(f"Created output directory {output_dir}")
+        visit_index = output_dir.parts.index(visit_name)
+        # Change permissions for folders in the visit directory and onwards
+        for current_path in list(reversed(output_dir.parents))[visit_index + 1 :]:
+            os.chmod(current_path, mode=machine_config.mkdir_chmod)
+
+    if PIL.Image is not None:
+        images = [PIL.Image.open(f) for f in gif_params.images]
+    else:
+        images = []
+    for im in images:
+        im.thumbnail((512, 512))
+
+    # Normalize and convert individual frames to 8-bit
+    arr: list[np.ndarray] = []
+    for im in images:
+        frame = np.array(im).astype(np.float32)
+        vmin, vmax = np.percentile(frame, (0.5, 99.5))
+        scale = 255 / ((vmax - vmin) or 1)
+        np.clip(frame, a_min=vmin, a_max=vmax, out=frame)
+        np.subtract(frame, vmin, out=frame)
+        np.multiply(frame, scale, out=frame)
+        arr.append(frame.astype(np.uint8))
+    arr = np.array(arr).astype(np.uint8)
+
+    # Convert back to PIL.Image objects and save as GIF
+    converted = [PIL.Image.fromarray(arr[f], mode="L") for f in range(len(images))]
+    converted[0].save(
+        gif_params.output_file,
+        format="GIF",
+        append_images=converted[1:],
+        save_all=True,
+        duration=30,
+        loop=0,
+    )
+
+    return {"output_gif": str(gif_params.output_file)}
